Refactor ImpressionCard.java
